requirements.txt
discord
pymysql
requests
pillow == 8.0.1